requirements.txt
Django 
djangorestframework
python-decouple
Pillow
drf_yasg
